Classic Mode
Classic Simon Says mode where the sequence of 50 numbers is pre
generated and the user just has to repeat the sequence adding one extra
pad to remember for each level.

Once the user reaches level 50 the game just continues currently.
Possibly add a congratulations message and a bonus round to this?